Create and use cluster list initializer
The cluster lists for regular files were only being initialized when the file contents were about to be read by Partition.read_file().  Unfortunately, this function only writes to a string buffer variable that stores all the file's contents - problematic for a multi-gigabyte file.  This could be better done with a generator, but that is an aside.  The DFXML loop wasn't using these file contents, so regular files weren't getting any clusters.

With this commit, the SHA-1 of "name.txt" now matches Fiwalk and uxtaf's SHA-1.